+++ CHANGELOG.md
@@ -9,8 +9,9 @@ GeoIP DB to avoid being blocked by security proxies
 - Add alignment booleans to JSON output
 - Fix `.msg` parsing CLI exception when `msgconvert` is not found in the 
 system path
-- Fix `SMTP AUTH extension not supported by server` error on some SMTP servers 
-(#12)
+- Add `--outgoing-port` and  `--outgoing-ssl` options
+- Fall back to plain text SMTP if `--outgoing-ssl` is not used and `STARTTLS` 
+is not supported by the server
 - Always use `\n` as the newline when generating CSVs
 
 3.9.7